+++ tools/build_production_image.sh
@@ -22,6 +22,8 @@ base_reference="$(jq -r '.base_image.reference' "${lock_path}")"
 base_digest="$(jq -r '.base_image.digest' "${lock_path}")"
 print_designer_url="$(jq -r '.apps.print_designer.url' "${lock_path}")"
 print_designer_commit="$(jq -r '.apps.print_designer.commit' "${lock_path}")"
+chromium_url="$(jq -r '.runtime_artifacts.print_designer_chromium.url' "${lock_path}")"
+chromium_sha256="$(jq -r '.runtime_artifacts.print_designer_chromium.sha256' "${lock_path}")"
 short_commit="$(git rev-parse --short=12 HEAD)"
 image_tag="${1:-erpnext-ua:${erpnext_ua_version}-${short_commit}}"
 
@@ -30,6 +32,8 @@ docker build \
 	--build-arg "ERPNEXT_UA_COMMIT=${erpnext_ua_commit}" \
 	--build-arg "PRINT_DESIGNER_COMMIT=${print_designer_commit}" \
 	--build-arg "PRINT_DESIGNER_URL=${print_designer_url}" \
+	--build-arg "CHROMIUM_URL=${chromium_url}" \
+	--build-arg "CHROMIUM_SHA256=${chromium_sha256}" \
 	--file "${containerfile_path}" \
 	--tag "${image_tag}" \
 	"${repo_root}"